chore(1-3389): move draggable item
